ui minor changes

+++ Ymir Engine/Source/G_UI.h
@@ -15,7 +15,7 @@ class UI_Canvas;
 class G_UI : public GameObject
 {
 public:
-	G_UI(UI_TYPE t, GameObject* pParent = External->scene->mRootNode, int x = 0, int y = 0);
+	G_UI(UI_TYPE t, GameObject* pParent = External->scene->mRootNode, float x = 0, float y = 0);
 	~G_UI();
 
 	update_status Update(float dt);
@@ -23,7 +23,7 @@ public:
 	std::vector<C_UI*> GetComponentsUI_ByType(UI_TYPE type);
 	//std::vector<C_UI*> GetComponentsUI_ByType(UI_TYPE type);
 
-	bool AddUIComponent(UI_TYPE type, int x, int y, GameObject* parent = External->scene->mRootNode);
+	bool AddUIComponent(UI_TYPE type, float x, float y, GameObject* parent = External->scene->mRootNode);
 
 public:
 	UI_Canvas* canvas;
