+++ Ymir Engine/Source/UI_Canvas.cpp
@@ -4,7 +4,7 @@
 
 #include "External/mmgr/mmgr.h"
 
-UI_Canvas::UI_Canvas(GameObject* g, int w, int h) : C_UI(UI_TYPE::CANVAS, ComponentType::UI, g, "Canvas")
+UI_Canvas::UI_Canvas(GameObject* g, float w, float h) : C_UI(UI_TYPE::CANVAS, ComponentType::UI, g, "Canvas")
 {
 	External->scene->vCanvas.push_back((G_UI*)mOwner);
 }
