+++ Ymir Engine/Source/UI_Canvas.h
@@ -8,7 +8,7 @@ class UI_Canvas : public C_UI
 {
 public:
 
-	UI_Canvas(GameObject* g, int w = 100, int h = 100);
+	UI_Canvas(GameObject* g, float w = 100, float h = 100);
 	~UI_Canvas();
 
 	void OnInspector();
